+++ R/hello.R
@@ -1,3 +1,0 @@
-hai <- function() {
-  print("Selamat Datang di Indonesia!")
-}