+++ R/spatial.R
@@ -10,33 +10,38 @@
 #' id_map("indonesia", "kota")
 #' id_map("jakarta", "rw")
 id_map <- function (region = "jakarta",
-                    level = "default") {
+                    level = "default",
+                    folder = system.file("data", "spatial", package = "nusantr")) {
 
   if (region == "jakarta") {
     if (level == "rw" | level == 5) {
-      reg <- readRDS(system.file("data", "spatial",
-                                 "jakarta_rw.rds", package = "indonesia"))
+      reg <- readRDS(sprintf("%s/%s", folder, "jakarta_rw.rds"))
     }
     else if (level == "default" | level == "kelurahan" | level == 4) {
-      reg <- readRDS(system.file("data", "spatial",
-                                 "jakarta_kelurahan.rds", package = "indonesia"))
+      reg <- readRDS(sprintf("%s/%s", folder, "jakarta_kelurahan.rds"))
     }
     else if (level == "kecamatan" | level == 3) {
-      reg <- readRDS(system.file("data", "spatial",
-                                 "jakarta_kecamatan.rds", package = "indonesia"))
+      reg <- readRDS(sprintf("%s/%s", folder, "jakarta_kecamatan.rds"))
     } else {
       stop("Error : for Jakarta region, level should be 'kecamatan', 'kelurahan' or 'rw'")
     }
   }
 
   else if (region == "indonesia") {
+    if (!file.exists(sprintf("%s/%s", folder, "indonesia_kota.rds"))) {
+      if (menu(c("Yes", "No"), title = "Do you want to download the map?") == 1) {
+        map_download()
+        map_convert()
+      } else {
+        stop("The map should be downloaded first")
+      }
+    }
+
     if (level == "kota" | level == 2) {
-      reg <- readRDS(system.file("data", "spatial",
-                                 "indonesia_kota.rds", package = "indonesia"))
+      reg <- readRDS(sprintf("%s/%s", folder, "indonesia_kota.rds"))
     }
     else if (level == "default" | level == "provinsi" | level == 1) {
-      reg <- readRDS(system.file("data", "spatial",
-                                 "indonesia_provinsi.rds", package = "indonesia"))
+      reg <- readRDS(sprintf("%s/%s", folder, "indonesia_provinsi.rds"))
     } else {
       stop("Error: for Indonesia region, level should be 'provinsi' or 'kota'")
     }